requirements.txt
asgiref==3.4.1
coverage==6.0.2
Django==3.2.8
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.3
sqlparse==0.4.2
